% Auto-generated: do not edit by hand
\name{dashDistributionIndicator}

\alias{dashDistributionIndicator}

\title{DashDistributionIndicator component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashDistributionIndicator(id=NULL, actualValue=NULL, colorActual=NULL,
colorExpected=NULL, colorInterval=NULL, expectedValue=NULL,
height=NULL, interval=NULL, style=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{actualValue}{Numeric. Actual value of the distribution}

\item{colorActual}{Character. The color of the actual value indicator. A string representation of hex value}

\item{colorExpected}{Character. The color of the expected value indicator. A string representation of hex value}

\item{colorInterval}{Unnamed list. Interval of five hex colors providing the color for the intervals}

\item{expectedValue}{Numeric. Expected value of the distribution}

\item{height}{Numeric. }

\item{interval}{Unnamed list. Intervals, an array of 6 values providing 5 intervals}

\item{style}{Unnamed list. Interval of five hex colors providing the color for the intervals}

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

